Move expr tests from test_bng to test_expressions
